重构数据库类Ⅰ
在这个提交中,我将需要用到数据库操作的方法抽取到 DAO 操作类中
并将这些方法改成公共的类方法,在 DAO 类中通过构造方法连接数据库
