+++ .gitignore
@@ -25,3 +25,4 @@ hs_err_pid*
 *.iml
 target/
 .idea/
+*.db
